+++ packages/categories/i18n/en.js
@@ -2,6 +2,9 @@
 i18n.map('en', {
 	title_cats_places: "Places by Category: {$1}",
 	title_cats_users: "Users by Category: {$1}",
+	
+	title_panelCatsType: "Manage Categories",
+
 	label_cat: "Category",
 	label_cats: "Categories",
 });