+++ packages/core/modules/filters.js
@@ -37,7 +37,7 @@ Kepler.filters = {
 //Places
 	placePanel: {
 		fields: {
-			loc:1, name:1, indoor:1, checkins:1, createdAt:1, userId:1, desc:1, warn:1, hist:1, source:1
+			loc:1, name:1, indoor:1, checkins:1, createdAt:1, userId:1, desc:1, warn:1, hist:1, url:1, source:1
 		}
 	},
 	placeItem: {