+++ packages/core/settings.js
@@ -73,6 +73,9 @@ Kepler.settings = {
 				"panelPlace_ui_btns": { order:-10 },
 				"panelPlace_ui_info": { order:-10 }
 			},
+			"tabPlace": {
+				"tabPlace_ui_info": { order:-10 }
+			},			
 			"panelUser": {
 				"panelUser_ui_bio": { order:-20 },
 				"panelUser_ui_status": { order:-20 },