+++ packages/osm/server/places.js
@@ -34,12 +34,14 @@ var osmToPlace = function(osm) {
 	var place = {
 		name: name, //K.Util.sanitize.name(name),
 		loc: [coords[1], coords[0]],
-		active: 0,
 		osm: feature,
 		source: {
 			type: 'osm'
 		}
 	};
+
+	if(prop.tags.website)
+		place.url = prop.tags.website;
 	
 	return place;
 };