+++ packages/core/modules/filters.js
@@ -7,7 +7,7 @@ Kepler.filters = {
 //Profile
 	currentUser: {
 		fields: {
-			name:1, username:1, avatar:1, usersBlocked:1, status:1, statusDefault:1, statusConnection:1, checkin:1, loc:1, mob:1, loginAt:1, createdAt:1, friends:1, gender:1, city:1, lang:1, loclast:1, hist:1, 
+			name:1, username:1, avatar:1, usersBlocked:1, status:1, statusDefault:1, statusConnection:1, checkin:1, loc:1, mob:1, loginAt:1, createdAt:1, friends:1, gender:1, city:1, lang:1, url:1, loclast:1, hist:1, 
 			
 			usersPending:1, usersReceive:1, emails:1, settings:1, source:1, profile:1
 		}
@@ -15,7 +15,7 @@ Kepler.filters = {
 //Users
 	friendPanel: {
 		fields: {
-			name:1, username:1, avatar:1, usersBlocked:1, status:1, statusDefault:1, statusConnection:1, checkin:1, loc:1, mob:1, loginAt:1, createdAt:1, friends:1, gender:1, city:1, lang:1, loclast:1, hist:1
+			name:1, username:1, avatar:1, usersBlocked:1, status:1, statusDefault:1, statusConnection:1, checkin:1, loc:1, mob:1, loginAt:1, createdAt:1, friends:1, gender:1, city:1, lang:1, url:1, loclast:1, hist:1
 		}
 	},
 	friendItem: {
@@ -25,7 +25,7 @@ Kepler.filters = {
 	},
 	userPanel: {
 		fields: {
-			name:1, username:1, avatar:1, usersBlocked:1, /*********************************************************************/ createdAt:1, friends:1, gender:1, city:1, lang:1
+			name:1, username:1, avatar:1, usersBlocked:1, /********************************************************************************/ createdAt:1, friends:1, gender:1, city:1, lang:1, url:1
 		}
 	},		
 	userItem: {