+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+...
+
+## v1.5.9
 - rename place file urls in urls
 - plugin edit url field is editable
 - plugin geoinfo fields calculate asyncronous on new place insert